requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
azure-ai-vision-imageanalysis==1.0.0b2
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.1
azure-storage-blob==12.20.0
backports.csv==1.0.7
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0
frozendict==2.4.4
frozenlist==1.4.1
greenlet==3.0.3
html5lib==1.1
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-core==0.2.1
langchain-text-splitters==0.2.0
langsmith==0.1.63
lxml==5.2.2
MarkupSafe==2.1.5
msrest==0.7.1
multidict==6.0.5
multitasking==0.0.11
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pdf2image==1.17.0
peewee==3.17.5
pillow==10.3.0
platformdirs==4.2.2
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==2.0.0
scipy==1.13.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
tablib==3.6.1
tenacity==8.3.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.3
Yahoo-ticker-downloader==3.0.1
yarl==1.9.4
yfinance==0.2.40
statsmodels==0.14.2
langchain-text_splitters==0.2.0